Added more tests and additional clenaup
