+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/MyDataSCU/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest.MyDataSCU
+{
+    public static class Constants
+    {
+        public const string CASHBOX_CERTIFICATION_ID = "";
+        public const string CASHBOX_CERTIFICATION_ACCESSTOKEN = "";
+    }
+}
