+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/MyDataSCU/Scenarios/AADECertificationExamples.cs
