+++ queue/test/fiskaltrust.Middleware.Localization.QueueGR.IntegrationTest/MyDataSCU/Scenarios/AADECertificationExamplesSelfPricing.cs
